refactor: Move expr formatter into its own crate
