+++ format/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "gluon_format"
+version = "0.5.0" # GLUON
+authors = ["Markus <marwes91@gmail.com>"]
+
+license = "MIT"
+
+description = "Code formatting for the gluon programming language"
+
+repository = "https://github.com/gluon-lang/gluon"
+documentation = "https://docs.rs/gluon"
+
+[dependencies]
+env_logger = { version = "0.3.4", optional = true }
+log = "0.3.6"
+pretty = "0.3.2"
+gluon_base = { path = "../base", version = "0.5.0" } # GLUON
+gluon_parser = { path = "../parser", version = "0.5.0" } # GLUON
+
+[dev-dependencies]
+difference = "1.0"
+pretty_assertions = "0.3.0"
+
+[features]
+test = ["env_logger"]
