+++ parser/src/lib.rs
@@ -420,30 +420,3 @@ pub fn parse_string<'env, 'input>(
 ) -> Result<SpannedExpr<String>, (Option<SpannedExpr<String>>, ParseErrors)> {
     parse_partial_expr(symbols, &TypeCache::new(), input)
 }
-
-pub fn format_expr(input: &str) -> Result<String, ParseErrors> {
-    use pretty::Arena;
-
-    use base::pretty_print::Printer;
-    use base::source::Source;
-    use base::symbol::Symbols;
-
-    let newline = match input.find(|c: char| c == '\n' || c == '\r') {
-        Some(i) => if input[i..].starts_with("\r\n") {
-            "\r\n"
-        } else if input[i..].starts_with("\r") {
-            "\r"
-        } else {
-            "\n"
-        },
-        None => "\n",
-    };
-
-    let type_cache = TypeCache::new();
-    let expr = parse_expr(&mut Symbols::new(), &type_cache, input)?;
-
-    let source = Source::new(input);
-    let arena = Arena::new();
-    let printer = Printer::new(&arena, &source);
-    Ok(printer.format(100, newline, &expr))
-}
