+++ repl/src/main.rs
@@ -12,6 +12,7 @@ extern crate log;
 extern crate gluon;
 #[macro_use]
 extern crate gluon_vm;
+extern crate gluon_format;
 
 use std::io::{self, Write};
 
@@ -19,6 +20,7 @@ use gluon::base;
 use gluon::parser;
 use gluon::check;
 use gluon::vm;
+use gluon::format;
 
 use base::error::InFile;
 
@@ -48,7 +50,7 @@ fn init_env_logger() {
 fn init_env_logger() {}
 
 fn format(writer: &mut Write, buffer: &str) -> Result<usize> {
-    use gluon::parser::format_expr;
+    use format::format_expr;
 
     let output = format_expr(buffer).map_err(|err| InFile::new("", buffer, err))?;
     writer.write_all(output.as_bytes())?;
